% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{.get_date}
\alias{.get_date}
\title{Helper function that is called by the exported function get_dates()}
\usage{
.get_date(date)
}
\arguments{
\item{date}{The single date for which data is to be queried}
}
\value{
A tibble containing all air quality data for that specific date
}
\description{
Helper function that is called by the exported function get_dates()
}
