% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.process_content}
\alias{.process_content}
\title{Helper function to transform API request content to tidy table format}
\usage{
.process_content(content)
}
\arguments{
\item{content}{The content to be tidied}
}
\value{
A tidied version of the API request content
}
\description{
Helper function to transform API request content to tidy table format
}
