% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{get_dates}
\alias{get_dates}
\title{Get all air quality data for a given date}
\usage{
get_dates(dates)
}
\arguments{
\item{dates}{A vector that is or can be coerced to a Date vector}
}
\value{
A tibble containing the data for all chosen dates
}
\description{
Get all air quality data for a given date
}
\examples{
my_date <- "2022-01-01"
d <- get_dates(my_date)

my_dates <- seq.Date(from = as.Date("2023-01-01"), to = as.Date("2023-01-11"), by = "days")
d <- get_dates(my_date)
}
