% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest.R
\name{get_latest}
\alias{get_latest}
\title{Get all air quality data from the last 24 hours}
\usage{
get_latest()
}
\value{
A tibble containing all air quality data that has been recorded in the last 24 hours
}
\description{
Get all air quality data from the last 24 hours
}
\examples{
d <- get_latest()
}
