% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station.R
\name{get_stations}
\alias{get_stations}
\title{Get info about all stations that have recorded any amount of data}
\usage{
get_stations()
}
\value{
A tibble containing basic information about the stations in the data
}
\description{
Get info about all stations that have recorded any amount of data
}
\examples{
stations <- get_stations()
}
