% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loftgaedi_2015}
\alias{loftgaedi_2015}
\title{Icelandic Air Quality Data}
\format{
\subsection{\code{who}}{

A data frame with 5,600,030 rows and 8 columns:
\describe{
\item{name}{Name of the station}
\item{local_id}{ID of the station}
\item{unit}{Measurement unit of observation}
\item{resolution}{Time-scale resolution of observation}
\item{verification}{Is the measurement verified or not}
\item{time}{Datetime of observation}
\item{variable}{Which variable is being measured?}
\item{value}{The value of the measurement}
...
}
}
}
\source{
\url{https://api.ust.is/aq}
}
\usage{
loftgaedi_2015
}
\description{
Air quality data in Iceland from 2015-01-01 to 2022-12-31
Report ...
}
\keyword{datasets}
